Закончена реализация приложения

+++ Organiser/Controllers/DailyPlannerController.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Organiser.Models;
 using Organiser.Models.Constants;
@@ -31,6 +30,9 @@ namespace Organiser.Controllers
 
         public ViewResult Edit(int eventId, string typeEvent)
         {
+            if (eventId <= 0 || string.IsNullOrEmpty(typeEvent))
+                return View("Error", "Что-то пошло не так");
+
             ViewBag.Title = "Редактирование";
             return View(eventsRepository.Events.FirstOrDefault(x => x.EventID == eventId && x.TypeGuid.ToString() == typeEvent));
         }
@@ -38,19 +40,22 @@ namespace Organiser.Controllers
         public ViewResult Create(string typeEvent)
         {
             ViewBag.Title = "Создание";
-            if (typeEvent == TypeGUIDs.CaseGuid.ToString())
+            if (typeEvent == Types.Case.Guid.ToString())
                 return View(nameof(Edit), new Case());
-            else if (typeEvent == TypeGUIDs.MeetGuid.ToString())
+            else if (typeEvent == Types.Meet.Guid.ToString())
                 return View(nameof(Edit), new Meet());
-            else if (typeEvent == TypeGUIDs.ReminderGuid.ToString())
+            else if (typeEvent == Types.Reminder.Guid.ToString())
                 return View(nameof(Edit), new Reminder());
             else
                 return View("Error", "Не найден тип!");
         }
 
         [HttpPost]
-        public RedirectResult Delete(int eventId, string typeEvent, string returnUrl)
+        public IActionResult Delete(int eventId, string typeEvent, string returnUrl)
         {
+            if (eventId <= 0 || string.IsNullOrEmpty(typeEvent) || string.IsNullOrEmpty(returnUrl))
+                return View("Error", "Что-то пошло не так");
+
             var deteleEntry = eventsRepository.DeleteEvent(eventId, typeEvent);
 
             if (deteleEntry != null)
@@ -77,6 +82,32 @@ namespace Organiser.Controllers
         public IActionResult EditReminder(Reminder reminder)
         {
             return EditEvent(reminder);
+        }
+
+        public ViewResult Filter([FromQuery(Name = "typeGuids")]List<string> typeGuids, string beginDate, string subject, string mode)
+        {
+            var events = eventsRepository.Events;
+
+            if (typeGuids != null && typeGuids.Any())
+            {
+                events = events.Where(x => typeGuids.Any(t => t == x.TypeGuid.ToString()));
+            }
+
+            if (!string.IsNullOrEmpty(beginDate) && DateTime.TryParse(beginDate, out DateTime date) && !string.IsNullOrEmpty(mode))
+            {
+                if (mode == Modes.Equally)
+                    events = events.Where(x => x.BeginDate == date);
+                if (mode == Modes.Less)
+                    events = events.Where(x => x.BeginDate <= date);
+                if (mode == Modes.More)
+                    events = events.Where(x => x.BeginDate >= date);
+            }
+
+            if (!string.IsNullOrEmpty(subject))
+            {
+                events = events.Where(x => x.Subject.Contains(subject));
+            }
+            return View(nameof(List), new EventsListModeShow { Events = events });
         }
 
         private IActionResult EditEvent(IEvent ev)
